Create Allones
This code describes a module called "allones" that takes an 8-bit input signal called "data" and outputs a 3-bit signal called "count". The purpose of the module is to count the number of "1"s in the input signal and output the count in binary format.

The module uses an "always" block with sensitivity to changes in the "data" input signal. Within the "always" block, the "count" signal is set equal to the 8-bit "data" signal with a leading 0 bit. This is accomplished using curly braces and the 1-bit value 0 followed by the 8 bits of the "data" signal. The count value is then incremented by 1 using the + operator and the 1-bit value 0. This effectively counts the number of 1s in the "data" signal.

Since the output signal is declared as a "reg", it can be assigned to within an "always" block, as is done in this code.